chore: 병합 후 저장
